Adding the touch-punch library for touch screens
This will allow us to use the jquery UI with the touch screens that we use for exhibits.

I had to move the script tags to the top of the html file since
touch-punch modifies the jquery ui at run time. To do this, the files need to be included in the head.