Rebuilt animation time handlers using tock.js
The setTimeout method wasn't very reliable on the looping animation
after it had run for awhile (8+ loops). The browser can be affected by
the computer's CPU, which causes latency in asychonous timers, so
tock.js queries the system clock on each loop, which is more reliable
for an animation like this one.